test loggers moved to android-test-common module, awaitSubscriberPresent added to PublisherAndSubscriberTests
